--- Shop.xcworkspace/xcuserdata/Moskwinow.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<Bucket
-   uuid = "04511B77-A6DB-4AD5-96D5-C903693570EF"
-   type = "0"
-   version = "2.0">
-</Bucket>
