--- Shop/App/BussinessLogic/Network/Login/Auth.swift
@@ -1,49 +1,0 @@
-//
-//  Auth.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Alamofire
-
-class Auth: AbstractRequestFactory {
-    let errorParser: AbstractErrorParser
-    let sessionManager: Session
-    let queue: DispatchQueue
-    let baseURL = URL(string: "https://raw.githubusercontent.com/GeekBrainsTutorial/online-store-api/master/responses/")!
-    
-    init(errorParser: AbstractErrorParser, sessionManager: Session, queue: DispatchQueue = DispatchQueue.global(qos: .utility)) {
-        self.errorParser = errorParser
-        self.sessionManager = sessionManager
-        self.queue = queue
-    }
-
-}
-
-extension Auth: AuthRequestFactory {
-    func login(userName: String, password: String, completionHandler: @escaping (AFDataResponse<LoginResult>) -> Void) {
-        let requestModel = Login(baseUrl: baseURL, login: userName, password: password)
-        self.request(request: requestModel, completionHandler: completionHandler)
-    }
-}
-
-extension Auth {
-    struct Login: RequestRouter {
-        let baseUrl: URL
-        let method: HTTPMethod = .get
-        let path: String = "login.json"
-        
-        let login: String
-        let password: String
-        var parameters: Parameters? {
-            return [
-                "username": login,
-                "password": password
-            ]
-        }
-    }
-}
-
-
