--- Shop/App/BussinessLogic/Network/Login/AuthRequestFactory.swift
@@ -1,14 +1,0 @@
-//
-//  AuthRequestFactory.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Alamofire
-
-protocol AuthRequestFactory {
-    func login(userName: String, password: String, completionHandler: @escaping(AFDataResponse<LoginResult>) -> ())
-    
-}
