--- Shop/App/BussinessLogic/Network/Update/UpdateUserFactory.swift
@@ -1,13 +1,0 @@
-//
-//  UpdateUserFactory.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Alamofire
-
-protocol UpdateUserFactory {
-     func updateUser(completionHandler: @escaping(AFDataResponse<UpdateUserResult>) -> ())
-}
