--- Shop/App/Entity/Logout/LogoutResult.swift
@@ -1,13 +1,0 @@
-//
-//  LogoutResult.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Foundation
-
-struct LogoutResult: Decodable {
-    let result: Int
-}
