--- Shop/App/Entity/Update/UpdateUserResult.swift
@@ -1,13 +1,0 @@
-//
-//  UpdateUserResult.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Foundation
-
-struct UpdateUserResult: Decodable {
-    let result: Int
-}
