--- Shop/App/Entity/User.swift
@@ -1,23 +1,0 @@
-//
-//  User.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Foundation
-
-struct User: Codable {
-    let id: Int
-    let login: String
-    let name: String
-    let lastname: String
-    
-    enum CodingKeys: String, CodingKey {
-        case id = "id_user"
-        case login = "user_login"
-        case name = "user_name"
-        case lastname = "user_lastname"
-    }
-}
