+++ Shop/Assets.xcassets/AppIcon.appiconset/Contents.json
