--- Shop/Core/Network/AbstractErrorParser.swift
@@ -1,14 +1,0 @@
-//
-//  AbstractErrorParser.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Foundation
-
-protocol AbstractErrorParser {
-    func parse(_ result: Error) -> Error
-    func parse(response: HTTPURLResponse?, data: Data?, error: Error?) -> Error?
-}
