--- Shop/Core/Network/ErrorParser.swift
@@ -1,22 +1,0 @@
-//
-//  ErrorParser.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Foundation
-
-class ErrorParser: AbstractErrorParser {
-    
-    func parse(_ result: Error) -> Error {
-        return result
-    }
-    
-    func parse(response: HTTPURLResponse?, data: Data?, error: Error?) -> Error? {
-        return error
-    }
-    
-    
-}
