--- Shop/Core/Network/RequestRouter.swift
@@ -1,44 +1,0 @@
-//
-//  RequestRouter.swift
-//  Shop
-//
-//  Created by Максим Вечирко on 20.12.2020.
-//  Copyright © 2020 Максим Вечирко. All rights reserved.
-//
-
-import Alamofire
-
-enum RequestRouterEncoding {
-    case url, json
-}
-
-protocol RequestRouter: URLRequestConvertible {
-    var baseUrl: URL {get}
-    var method: HTTPMethod {get}
-    var path: String {get}
-    var parameters: Parameters? {get}
-    var fullUrl: URL {get}
-    var encoding: RequestRouterEncoding {get}
-}
-
-extension RequestRouter {
-    
-    var fullUrl: URL {
-        return baseUrl.appendingPathComponent(path)
-    }
-    var encoding: RequestRouterEncoding {
-        return .url
-    }
-    
-     func asURLRequest() throws -> URLRequest {
-        var urlRequest = URLRequest(url: fullUrl)
-        urlRequest.httpMethod = method.rawValue
-        
-        switch self.encoding {
-        case .url: return try URLEncoding.default.encode(urlRequest, with: parameters)
-        case .json: return try JSONEncoding.default.encode(urlRequest, with: parameters)
-        
-        }
-    }
-    
-}
