+++ Shop/ViewController.swift
@@ -1,0 +1,20 @@
+//
+//  ViewController.swift
+//  Shop
+//
+//  Created by Максим Вечирко on 18.12.2020.
+//  Copyright © 2020 Максим Вечирко. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view.
+    }
+
+
+}
+
